changes
restructured the commands a bit to make it easier to find which one you want to work on. Hacked together a crude but workable directory awareness for the jobs. Can read results back from the dispatcher. We're close to actually being able to build some logic around the project video file.